Brings feat/media-engine-port up to the latest developer. The port replaces
the C media player with a Rust one, so the media player keeps its Rust side
throughout and takes none of developer's C engine:

- Every media-player .cs file resolved to the port's version. Developer's
  C-player resync (ResyncLocal/ResyncEveryone/ReloadInPlace, LoadApprovedUrl)
  is not taken; the Rust side already has Resync Everyone and does URL
  approval in the shims.
- BasisMediaPlayerNetworkingInspector and BasisCaptionCue auto-merged from
  developer and were overridden back to the port: developer's inspector
  wired a button to ResyncLocal, which the Rust networking does not have,
  and the caption-cue auto-merge had dropped the port's HasText.
- BasisVideoFrameBufferPool stays deleted (the Rust engine owns the pool).
- en.json unions both sides' new keys (the port's media-settings strings and
  developer's Quest controller-fix string).
- README and TESTING kept as the port's Rust-era docs.

Developer's non-media changes and its trusted-URL data (VRCDN, RTSP defaults)
come across normally. Deferred, not lost: developer's ResyncLocal asks the
room before reloading; the Rust side has a simpler local re-open. See
com.basis.mediaplayer/MERGE-NOTES.md.

+++ Basis/Packages/com.basis.server/BasisNetworkCore/Compression/BasisBoneRotationCompression.cs
@@ -622,7 +622,7 @@ namespace Basis.Network.Core.Compression
             uint qA = Clamp((uint)Math.Round((ClampF(b * invRange, -1f, 1f) * 0.5f + 0.5f) * maxQ), 0, maxQ);
             uint qC = Clamp((uint)Math.Round((ClampF(c * invRange, -1f, 1f) * 0.5f + 0.5f) * maxQ), 0, maxQ);
 
-            return (ulong)maxIdx | ((ulong)qa << 2) | ((ulong)qA << (2 + bpc)) | ((ulong)qC << (2 + 2 * bpc));
+            return (ulong)(uint)maxIdx | ((ulong)qa << 2) | ((ulong)qA << (2 + bpc)) | ((ulong)qC << (2 + 2 * bpc));
         }
 
         /// <summary>
